requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.8.2,<3.9.0
psycopg2>=2.7.5,<2.8.0

djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.6.1
pytz==2018.3
